Adding authorization password to save and delete laptop

+++ src/controllers/laptop.controller.ts
@@ -3,6 +3,8 @@ import { laptopServices } from "../services/laptop.services.js";
 import { DeleteLaptop, GetCartNameByLaptopCodeParam, GetLaptopsByCartIdParams, SaveLaptopInCartBody, UpdateLaptopCartBody } from "../resource/laptop.resource";
 import { createErrorResponse } from "../common/error.resource.js";
 import { HttpStatus } from "../common/httpStatus.js";
+import { verifyAdminPassword } from "../common/verifyAdminPassword.js";
+import { authController } from "./auth.controller.js";
 
 async function getAllLaptopsFromSpecificCart(request: FastifyRequest<{ Params: GetLaptopsByCartIdParams }>, reply: FastifyReply) {
     const { cartSlug } = request.params
@@ -29,7 +31,14 @@ async function getCartNameByLaptopCode(request: FastifyRequest<{ Params: GetCart
 }
 
 async function saveLaptop(request: FastifyRequest<{ Body: SaveLaptopInCartBody }>, reply: FastifyReply) {
-    const { laptopCode, cartSlug } = request.body;
+    const { laptopCode, cartSlug, password } = request.body;
+    const hasAccess = authController.checkUser(password)
+    
+    
+    if (!hasAccess) {
+        return reply.status(HttpStatus.UNAUTHORIZED).send(createErrorResponse(`Senha Incorreta`))
+    }
+    
     const laptop = await laptopServices.saveLaptop(laptopCode, cartSlug)
 
     if (laptop.isError()) {
@@ -49,12 +58,26 @@ async function updateLaptopCart(request: FastifyRequest<{ Body: UpdateLaptopCart
 }
 
 async function deleteLaptopCart(request: FastifyRequest<{ Body: DeleteLaptop }>, reply: FastifyReply) {
-    const { laptopCode } = request.body
+    const { laptopCode, password } = request.body
 
-    const laptop = await laptopServices.deleteLaptop(laptopCode)
+    const hasAccess = authController.checkUser(password)
 
-    if (laptop.isError()) {
-        return reply.status(HttpStatus.NOT_FOUND).send(createErrorResponse(`${laptop.error}`))
+    if (!hasAccess) {
+        return reply.status(HttpStatus.UNAUTHORIZED).send(createErrorResponse(`Senha Incorreta`))
+    }
+
+    const laptop = await laptopServices.laptopExists(laptopCode)
+
+    if (!laptop) {
+        return reply
+            .status(HttpStatus.NOT_FOUND)
+            .send(createErrorResponse("Notebook não encontrado"))
+    }
+
+    const deletedLaptop = await laptopServices.deleteLaptop(laptopCode)
+
+    if (deletedLaptop.isError()) {
+        return reply.status(HttpStatus.NOT_FOUND).send(createErrorResponse(`${deletedLaptop.error}`))
     }
     return reply.status(HttpStatus.OK).send(laptop)
 }
